requirements.txt
Flask
jsonify
request
make_response
pyjwt